Compatibility with SCS
SCS includes a component that makes contingencies and spell triggers innate abilities and, as such, removes all of those scrolls from the game. Since a contingency scroll is required for the various archmagi robe upgrades, Cespenar and Cromwell will instead adjust the formula to use a scroll of Globe of Invulnerability instead.

I will also be sending a pull request to David for SCS--as it should be installed after IU, it should detect and add a small bit of scripting for Cromwell and Cespenar to adjust their formulas accordingly.